// src/components/dynamicForm/types.ts
export interface FormRenderConfigType {
  ordering: "auto" | "sequence";
  renderType: "simple" | "tabs" | "stepper" | "accordian"; //implimentation pending
  labels?: {
    next?: string;
    prev?: string;
    complete?: string;
  };
  groups?: { [key: string]: string };
  gridConfig: {
    item?: {
      xs?: GridSize;
      sm?: GridSize;
      md?: GridSize;
      spacing?: GridSpacing;
    };
    container: {
      direction?: GridDirection;
      spacing?: GridSpacing;
    };
  };
}

export interface UserFlowType {
  componentName: string;
  sequence: number;
}

export interface FormMetaDataType {
  name: string;
  label: string;
  resetFieldOnUnmount: boolean;
  validationRun: "onBlur" | "onChange" | "all";
  render: FormRenderConfigType;
  componentProps: ComponentTypeProps;
  formState?: any;
  //very specific need better api and remove from metaData -future TODO
  flow?: UserFlowType[];
  submitAction?: string;
  refID?: string | number; //remove it and make it part of formState
  confirmationBox?: {
    name: string;
    label: string;
  };
}

export interface ComponentTypeProps {
  textField?: TextFieldPropsOptional;
  select?: SelectPropsOptional;
  checkbox?: CheckboxPropsOptional;
  checkboxGroup?: CheckboxGroupPropsOptional;
  datetimePicker?: DateTimePickerPropsOptional;
  datePicker?: DatePickerPropsOptional;
  timePicker?: TimePickerPropsOptional;
  radio?: RadioPropsOptional;
  rating?: RatingPropsOptional;
  slider?: SliderPropsOptional;
  switch?: SwitchPropsOptional;
  switchGroup?: SwitchGroupPropsOptional;
  numberFormat?: NumberFormatPropsOptional;
  passwordField?: PasswordFieldPropsOptional;
  toggleButtonGroup?: ToggleButtonGroupPropsOptional;
  inputMask?: InputMaskPropsOptional;
  autocomplete?: AutocompletePropsOptional;
  textarea?: TextareaAutosizeFieldPropsOptional;
  searchField?: SearchFieldPropsOptional;
  visaversa?: VisaversaFieldPropsOptional;
  dataTable?: DataTablePropsOptional;
}

export interface MetaDataType {
  form: FormMetaDataType;
  fields: FieldMetaDataType[];
}

//when you change a type _fields here also change it in common ArrayFields component

export type FieldMetaDataTypeX =
  | AllTextFieldProps
  | AllSelectFieldProps
  | AllCheckboxGroupProps
  | AllCheckboxProps
  | AllDatePickerProps
  | AllDateTimePickerProps
  | AllTimePickerProps
  | AllRadioProps
  | AllRatingProps
  | AllSliderProps
  | AllSwitchGroupProps
  | AllSwitchProps
  | AllNumberFormatProps
  | AllPasswordFieldProps
  | AllToggleButtonGroupProps
  | AllSpacerProps
  | AllInputMaskProps
  | AllAutocompleteProps
  | AllTextareaAutosizeFieldProps
  | AllTypographyFieldProps
  | AllHiddenFieldProps
  | AllTransferListProps
  | AllSearchFieldProps
  | AllVisaversaFieldProps
  | ArrayFieldProps
  | AllDataTableProps
  | AllFormButtonProps
  | AllAccountNumberProps
  | AllLoginIDProps;

export type FieldMetaDataType = Merge<
  FieldMetaDataTypeX,
  { template?: FieldMetaDataType[] }
>;
// ...
export type FieldMetaDataTypeOptional =
  | TextFieldPropsOptional
  | SelectPropsOptional
  | CheckboxPropsOptional
  | CheckboxGroupPropsOptional
  | DateTimePickerPropsOptional
  | DatePickerPropsOptional
  | TimePickerPropsOptional
  | RadioPropsOptional
  | RatingPropsOptional
  | SliderPropsOptional
  | SwitchPropsOptional
  | SwitchGroupPropsOptional
  | NumberFormatPropsOptional
  | PasswordFieldPropsOptional
  | ToggleButtonGroupPropsOptional
  | InputMaskPropsOptional
  | AutocompletePropsOptional
  | TextareaAutosizeFieldPropsOptional
  | SearchFieldPropsOptional
  | VisaversaFieldPropsOptional
  | DataTablePropsOptional;

export type ExtendedFieldMetaDataTypeOptional = {
  [key: string]: FieldMetaDataTypeOptional;
};